create grid scaffolding
