+++ src/components/atoms/GameGridCell/styled.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const GameGridCellContainer = styled.div``;
