+++ src/components/organisms/GameGrid/styled.ts
